[ feat ] 구조 변경
